backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.7.1
pydantic-settings==2.4.0
python-multipart==0.0.6
pyjwt==2.8.0
bcrypt==4.1.2
pytest==7.4.3
volcengine-python-sdk[ark]==1.0.101